requirements.txt
streamlit>=1.28.0
crewai>=0.28.0
sqlalchemy>=2.0.0
pyodbc>=4.0.39
python-dotenv>=1.0.0
pandas>=2.0.0
openai>=1.0.0
plotly>=5.17.0
